mod header to have links as a navbar, added sign in and sign up components
