+++ .env
@@ -1,0 +1,4 @@
+REACT_APP_PROVIDER_KEY_EMAIL=56pbDNOK9kQrevm8Em1CShXbY093
+REACT_APP_PROVIDER_KEY_FB=
+REACT_APP_PROVIDER_KEY_GOOGLE=
+REACT_APP_PROVIDER_KEY_GH=
