+++ package.json
@@ -16,6 +16,7 @@
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-redux": "^7.2.6",
+    "react-router-dom": "^6.9.0",
     "react-scripts": "5.0.0",
     "reselect": "^4.1.5",
     "styled-components": "^5.3.3",
