+++ src/components/LogIn.jsx
@@ -1,0 +1,5 @@
+import React from "react";
+
+const LogIn = () => {};
+
+export default LogIn;
