+++ src/firebase/auth.js
@@ -1,0 +1,39 @@
+const firebaseui = require("firebaseui");
+var firebase = require("firebase");
+
+var ui = new firebaseui.auth.AuthUI(firebase.auth());
+
+// provider id = 56pbDNOK9kQrevm8Em1CShXbY093
+
+const uiConfig = {
+  callbacks: {
+    signInSuccessWithAuthResult: function (authResult, redirectUrl) {
+      // User successfully signed in.
+      // Return type determines whether we continue the redirect automatically
+      // or whether we leave that to developer to handle.
+      return true;
+    },
+    uiShown: function () {
+      // The widget is rendered.
+      // Hide the loader.
+      document.getElementById("loader").style.display = "none";
+    },
+  },
+  // Will use popup for IDP Providers sign-in flow instead of the default, redirect.
+  signInFlow: "popup",
+  signInSuccessUrl: "/",
+  signInOptions: [
+    // Leave the lines as is for the providers you want to offer your users.
+    // firebase.auth.GoogleAuthProvider.PROVIDER_ID,
+    // firebase.auth.FacebookAuthProvider.PROVIDER_ID,
+    // firebase.auth.GithubAuthProvider.PROVIDER_ID,
+    firebase.auth.EmailAuthProvider.PROVIDER_ID,
+    // firebase.auth.PhoneAuthProvider.PROVIDER_ID,
+  ],
+  // Terms of service url.
+  // tosUrl: "<your-tos-url>",
+  // Privacy policy url.
+  // privacyPolicyUrl: "<your-privacy-policy-url>",
+};
+
+export const auth = ui.start("#firebaseui-auth-container", uiConfig);
